baby-cry-ex built

+++ ch12/baby_monitor/edge-impulse-sdk/CMSIS/DSP/Source/FastMathFunctions/arm_atan2_f16.c
@@ -137,7 +137,7 @@ arm_status arm_atan2_f16(float16_t y,float16_t x,float16_t *result)
         }
         else
         {
-            if (signbit(y))
+            if (signbit((float)y))
             {
                *result= -(_Float16)PIF16;
             }
